+++ src/billiards/defs_animating.jl
@@ -1,195 +1,197 @@
 using DataStructures: CircularBuffer
+######################################################################################
+# Code outline
+######################################################################################
+BILLIARDS_ANIMATIONS_API = """
+## Description
+Expected inputs: vector of particles, a billiard, and optionally the output of
+`arcintervals`, which initializes plotted elements related with the boundary map
 
-# TODO: I Think I need to scrap this entire design,
-# and make a new design based on a Observable{Vector{Particles}}...
-# then, everything can be lift/on of this observable. The stepping function
-# would bounce particles and then just call `notify`. All else gets updated
-# for free. Also, all stuff should be vectors of values, each for each particle.
+Two helper structures for each particle:
+1. `ParticleHelper`: Contains quantities that are updated each `dt` step,
+   including the particle and its tail.
+2. `CollisionHelper`: Contains quantities that are only updated each collision.
+
+These two helpers are necessary to
+split the simulation into real-time stepping, instead of the traditional
+DynamicalBilliards.jl setup of discrete time stepping (1 step = 1 collision).
+
+The main thing returned to the user are two observables, one having vector
+of helpers (1) and the other having vector of helpers (2).
+Every plotted element is lifted from these observables.
+
+An exported high-level function `bdplot_animstep!(phs, chs, bd, dt; update, intervals)`
+progresses the simulation for one `dt` step. The user should be using this function for
+custom-made animations. The only thing the `update` keyword does is:
+`notify!(phs)`. The collision helper is always notified at collisions.
+"""
 
 ######################################################################################
-# Observable for single particle, which auto-steps its plotted tail/boundary map
+# Struct definitions
 ######################################################################################
-
-mutable struct ParticleObservable{T<:Real, P<:AbstractParticle}
-    # Fields necessary for simulation
-    p::P         # particle
-    i::Int       # index of obstacle to be collided with
-    tmin::T      # time to next collision
-    t::T         # current time (resets at each collision)
-    n::Int       # number of collisions done so far
-    T::T         # total time
-    # Fields used in plotting
-    tail::Observable{CircularBuffer{Point2f}}
-    ξsin::Observable{Point2f}   # Only used when plotting in boundary map
+mutable struct ParticleHelper{P, T<:Real}
+    p::P           # particle
+    t::T           # time ellapsed (resets at each collision)
+    T::T           # total time ellapsed (only resets when resetting particles)
+    tail::CircularBuffer{Point2f} # particle positions in last recorded steps
 end
-const ParObs = ParticleObservable
 
-function ParticleObservable(p::P, bd, n, ξsin = Point2f(0, 0)) where {P<:AbstractParticle}
+mutable struct CollisionHelper{T<:Real}
+    i::Int         # index of obstacle to be collided with
+    tmin::T        # time to next collision
+    n::Int         # total collisions so far
+    ξsinφ::SVector{2, T}  # boundary map point
+end
+
+function helpers_from_particle(p::AbstractParticle, bd::Billiard, L, intervals = nothing)
     T = eltype(p.pos)
-    i, tmin, cp = DynamicalBilliards.next_collision(p, bd)
-    cb = CircularBuffer{Point2f}(n)
-    fill!(cb, Point2f(p.pos))
-    ParticleObservable{T,P}(p, i, tmin, 0, 0, 0, Observable(cb), Observable(ξsin))
-end
-
-function rebind_partobs!(p::ParticleObservable, p0::AbstractParticle, bd, ξsin = p.ξsin[])
-    i, tmin, cp = DynamicalBilliards.next_collision(p0, bd)
-    ξ = sφ = 0f0 # TODO: Use boundary map on cp
-    p.p.pos = p0.pos
-    p.p.vel = p0.vel
-    DynamicalBilliards.ismagnetic(p.p) && (p.p.center = DynamicalBilliards.find_cyclotron(p.p))
-    p.i, p.tmin, p.t, p.n, p.T = i, tmin, 0, 0, 0
-    L = length(p.tail[])
-    append!(p.tail[], [Point2f(p0.pos) for i in 1:L])
-    p.tail[] = p.tail[]
-    if ξsin !== nothing
-        p.ξsin[] = ξsin # This can only be updated from bmap, which gives selection directly
-    end
-end
-
-function billiard_animstep!(parobs::ParticleObservable, bd, dt, updateplot = true, intervals = nothing)
-    if parobs.t + dt - parobs.tmin > 0
-        # There is a collision happening within this window,
-        # and hence we need to update the stuff
-        rt = parobs.tmin - parobs.t # remaining time
-        animbounce!(parobs, bd, rt, updateplot, intervals)
+    i, tmin, _ = DynamicalBilliards.next_collision(p, bd)
+    if !isnothing(intervals)
+        p2 = propagate!(copy(p), tmin)
+        ξ, sinφ = to_bcoords(p2.pos, p2.vel, bd[i])
+        ξsin = SVector(ξ, sinφ)
     else
-        DynamicalBilliards.propagate!(parobs.p, dt)
-        parobs.t += dt
-        push!(parobs.tail[], parobs.p.pos)
-        if updateplot
-            parobs.tail[] = parobs.tail[] # trigger update
-        end
+        ξsin = SVector{2, T}(0, 0)
     end
+    tail = CircularBuffer{Point2f}(L)
+    fill!(tail, Point2f(p.pos))
+    return ParticleHelper{typeof(p), T}(p, 0, 0, tail), CollisionHelper{T}(i, tmin, 0, ξsin)
+end
+
+function helpers_from_particles(
+    ps::Vector{P}, bd::Billiard, L, intervals = nothing) where {P<:AbstractParticle}
+    T = eltype(ps[1])
+    phs = ParticleHelper{P, T}[]
+    chs = CollisionHelper{T}[]
+    for i in 1:length(ps)
+        ph, ch = helpers_from_particle(ps[i], bd, L, intervals)
+        push!(phs, ph); push!(chs, ch)
+    end
+    return Observable(phs), Observable(chs)
+end
+
+
+
+######################################################################################
+# Stepping functions
+######################################################################################
+function bdplot_animstep!(phs, chs, bd, dt; update = false, intervals = nothing)
+    phs_val = phs[]; chs_val = chs[]
+    N = length(phs_val)
+    any_collided = false
+    for i in 1:N
+        collided = bdplot_animstep!(phs_val[i], chs_val[i], bd, dt, intervals)
+        any_collided = collided || any_collided
+    end
+    any_collided && notify(chs)
+    update && notify(phs)
     return
 end
 
-function animbounce!(parobs, bd, rt, updateplot = true, intervals = nothing)
-    DynamicalBilliards.propagate!(parobs.p, rt)
-    DynamicalBilliards._correct_pos!(parobs.p, bd[parobs.i])
-    DynamicalBilliards.resolvecollision!(parobs.p, bd[parobs.i])
-    DynamicalBilliards.ismagnetic(parobs.p) && (parobs.p.center = DynamicalBilliards.find_cyclotron(parobs.p))
-    # `intervals` are the boundary map intervals (needs knowledge of DynamicalBilliards.jl)
-    if intervals !== nothing
-        ξ, sφ = DynamicalBilliards.to_bcoords(parobs.p.pos, parobs.p.vel, bd[parobs.i])
-        ξ += intervals[parobs.i]
-        parobs.ξsin[] = (ξ, sφ)
+function bdplot_animstep!(ph::ParticleHelper, ch::CollisionHelper, bd, dt, intervals)
+    collided = false
+    if ph.t + dt - ch.tmin > 0
+        # We reach the collision point within the `dt` window, so we need to
+        # call the "bounce" logic
+        rt = ch.tmin - ph.t # remaining time to collision
+        billiards_animbounce!(ph, ch, bd, rt, intervals)
+        dt = dt - rt # remaining dt to propagate for
+        collided = true
     end
-    i, tmin, = DynamicalBilliards.next_collision(parobs.p, bd)
-    parobs.i = i
-    parobs.tmin = tmin
-    parobs.t = 0
-    parobs.T += tmin
-    parobs.n += 1
-    push!(parobs.tail[], parobs.p.pos)
-    if updateplot
-        parobs.tail[] = parobs.tail[] # trigger update
+    DynamicalBilliards.propagate!(ph.p, dt)
+    ph.t += dt
+    ph.T += dt
+    push!(ph.tail, ph.p.pos)
+    return collided
+end
+
+function billiards_animbounce!(ph::ParticleHelper, ch::CollisionHelper, bd, rt, intervals)
+    # Bring to collision and resolve it:
+    DynamicalBilliards.propagate!(ph.p, rt)
+    DynamicalBilliards._correct_pos!(ph.p, bd[ch.i])
+    DynamicalBilliards.resolvecollision!(ph.p, bd[ch.i])
+    DynamicalBilliards.ismagnetic(ph.p) && (ph.p.center = DynamicalBilliards.find_cyclotron(ph.p))
+    # Update all remaining counters
+    i, tmin, = DynamicalBilliards.next_collision(ph.p, bd)
+    ch.i = i
+    ch.tmin = tmin
+    ch.n += 1
+    ph.t = 0
+    ph.T += rt
+    # Compute new boundary map point if necessary
+    if !isnothing(intervals)
+        ξ, sφ = DynamicalBilliards.to_bcoords(ph.p.pos, ph.p.vel, bd[parobs.i])
+        ξ += intervals[ph.i]
+        ch.ξsin = SVector(ξ, sφ)
     end
+    # Also add the collision point to the tail for continuity
+    push!(ph.tail, ph.p.pos)
     return
 end
 
 
 ######################################################################################
-# Animation stepper for a group of particles. Includes plotted quiver field
+# Initialization of observables and their plots
 ######################################################################################
-
-mutable struct ParticleStepper{T<:Real, P<:AbstractParticle, X}
-    allparobs::Vector{ParticleObservable{T, P}} # individual particles
-    balls::Observable{Vector{Point2f}}          # data for plotting all particles
-    vels::Observable{Vector{Point2f}}           # data for plotting all particles
-    update_balls::Bool                          # whether to update balls
-    vellength::Float32                          # scaling of velocity vector
-    bmap_points::X # scatter points for boundary map, vector{observable{vector{point}}}
-    N::Int                                      # convenience, = length(allparobs)
-end
-
-function bdplot_initialize_stepper!(ax, ps::Vector{<:AbstractParticle}, bd;
-        # Remaining arguments for tuning plotting, e.g. color, linewidths, etc.
-        α = 1.0, colors = JULIADYNAMICS_CMAP, fade = true, tailwidth = 1,
-        # Keywords for plotting the particles
+function bdplot_plotting_init!(ax::Axis, bd::Billiard, ps::Vector{<:AbstractParticle};
+        tail_length = 1000, colors = JULIADYNAMICS_CMAP,
+        fade = true,
+        tailwidth = 1,
         plot_particles = true,
-        arrowsize = 1, particlewidth = 3tailwidth, particlesize = 1, vellength = 2.0,
-        # boundary map arguments
-        axbmap = nothing,
+        particle_size = 5.0, # size of marker for scatter plot of particle balls
+        velocity_size = 0.05, # size of multiplying the quiver
+        bmax = nothing, intervals = nothing,
     )
 
+    bdplot!(ax, bd)
     N = length(ps)
-    allparobs = [ParObs(p, bd, tail) for p in ps]
     cs = if !(colors isa Vector) || length(colors) ≠ N
-        colors_from_map(colors, α, N)
+        InteractiveDynamics.colors_from_map(colors, N, 0.8)
     else
         to_color.(colors)
     end
+    # Instantiate the helper observables
+    phs, chs = helpers_from_particles(ps, bd, tail_length)
 
+    ######################################################################################
+    # Initialize plot elements and link them via Observable pipeline
+    # Tail circular data:
+    tails = [Observable(p.tail) for p in phs[]]
     # Plot tails
-    for (i, p) in enumerate(allparobs)
+    for i in 1:N
         x = to_color(cs[i])
         if fade
-            x = [RGBAf0(x.r, x.g, x.b, i/tail) for i in 1:tail]
+            x = [RGBAf(x.r, x.g, x.b, i/tail_length) for i in 1:tail_length]
         end
-        lines!(ax, p.tail; color = x, linewidth = tailwidth)
+        lines!(ax, tails[i]; color = x, linewidth = tailwidth)
+    end
+    # Trigger tail updates (we need `on`, can't use `lift`, coz of `push!` into buffer)
+    on(phs) do phs
+        for i in 1:N
+            tails[i][] = phs[i].tail
+            notify(tails[i])
+        end
     end
 
-    balls = Observable([Point2f(p.p.pos) for p in allparobs])
-    vels = Observable([vellength*Point2f(p.p.vel) for p in allparobs])
+    # Particles and their quiver
     if plot_particles # plot ball and arrow as a particle
+        balls = lift(phs -> [Point2f(ph.p.pos) for ph in phs], phs)
+        vels = lift(phs -> [Point2f(velocity_size*ph.p.vel) for ph in phs], phs)
         scatter!(
             ax, balls; color = darken_color.(cs),
-            marker = MARKER, markersize = 8*particlesize*Makie.px,
-            strokewidth = 0.0,
-        ),
+            markersize = particle_size, strokewidth = 0.0,
+        )
         arrows!(
             ax, balls, vels; arrowcolor = darken_color.(cs),
             linecolor = darken_color.(cs),
-            arrowsize = arrowsize,
-            linewidth  = particlewidth*4,
+            normalize = false,
+            # arrowsize = particle_size*vr/3,
+            linewidth  = 2,
         )
     end
 
-    # TODO: Boundary map
-    # probably `lift` the `balls`...?
-    if !isnothing(axbmap)
-        bmap_points = [Observable([Point2f(p[].ξsin)]) for p in allparobs]
-        for i in 1:length(ps)
-            parobs = allparobs[i]
-            on(parobs.ξsin) do val
-                push!(bmap_points[i], val)
-            end
-        end
-    end
+    # TODO: boundary map observable lifting and plotting
 
-
-
-
-    return ParticleStepper(allparobs, balls, vels, plot_particles, vellength, length(ps))
+    return phs, chs
 end
 
-function billiard_animstep!(stepper::ParticleStepper, bd, dt, n, intervals = nothing)
-    for _ in 1:n
-        for i in 1:stepper.N
-            parobs = stepper.allparobs[i]
-            billiard_animstep!(parobs, bd, dt, false, intervals)
-        end
-    end
-
-    for i in 1:stepper.N
-        parobs = stepper.allparobs[i]
-        billiard_animstep!(parobs, bd, dt, false, intervals)
-        if stepper.update_balls
-            balls[][i] = parobs.p.pos
-            vels[][i] = stepper.vellength * parobs.p.vel
-        end
-    end
-    if stepper.update_balls
-        balls[] = balls[]
-        vels[] = vels[]
-    end
-    # TODO: Boundary map here.
-
-    if !isnothing(intervals)
-        fafa
-    end
-end
-
-
-# TODO: Rebind stepper function, from `interactive_billiard`.
