+++ src/billiards/exports.jl
@@ -1,5 +1,6 @@
 export bdplot, bdplot!
+export bdplot_animstep!
 export interactive_billiard
 export interactive_billiard_bmap
-export billiard_video 
+export billiard_video
 export billiard_video_timeseries
